Move logic of the index file to App.js and adjust tests accordingly
