+++ js/__tests__/App.spec.js
@@ -4,11 +4,11 @@ import React from 'react';
 import Renderer from 'react-test-renderer';
 
 // Components
-import Index from '../../index';
+import App from '../App';
 
 
 it('renders correctly', () => {
   const tree = Renderer.create(
-    <Index />
+    <App />
   );
 });
